Renamed discovery route to discover; Minor changes
