+++ src/components/Discover/Card.tsx
