+++ src/components/Discover/ContentModeBtn.tsx
